+++ src/app/core/home/home.component.ts
@@ -21,7 +21,7 @@ export class HomeComponent implements AfterViewInit {
   public fullHomeScreen = true;
   public showFullCyanHomeBtn: Boolean = true;
   public selectedColor = 'rgb(255, 255, 255)';
-  public homeColor = '#f2e6b1';
+  public homeColor = '#f2ccb1';
   public infoColor = '#95dab6';
   public mapColor = '#83b2d0';
   public graphColor = '#7f87b2';
@@ -58,6 +58,7 @@ export class HomeComponent implements AfterViewInit {
     this.showGraph = false;
     this.showIntro = true;
     this.selectBtn('home');
+    this.componentDisplayService.getDisableMap(true);
   }
 
   public clickMap() {
@@ -66,6 +67,7 @@ export class HomeComponent implements AfterViewInit {
     this.showGraph = false;
     this.showIntro = false;
     this.selectBtn('map');
+    this.componentDisplayService.getDisableMap(false);
   }
 
   public clickInfo() {
@@ -74,6 +76,7 @@ export class HomeComponent implements AfterViewInit {
     this.showGraph = false;
     this.showIntro = false;
     this.selectBtn('info');
+    this.componentDisplayService.getDisableMap(true);
   }
 
   public clickGraph() {
@@ -82,6 +85,7 @@ export class HomeComponent implements AfterViewInit {
     this.showMap = false;
     this.showIntro = false;
     this.selectBtn('graph');
+    this.componentDisplayService.getDisableMap(true);
   }
 
   public selectBtn(button) {
@@ -98,9 +102,9 @@ export class HomeComponent implements AfterViewInit {
     }
     if (button !== 'home') {
       homeBtnFullID.style.color = 'rgb(255, 255, 255)';
-      homeBtnFullID.style.backgroundColor = '#f2e6b1';
+      homeBtnFullID.style.backgroundColor = this.homeColor;
       homeBtnSmID.style.color = 'rgb(255, 255, 255)';
-      homeBtnSmID.style.backgroundColor = '#f2e6b1';
+      homeBtnSmID.style.backgroundColor = this.homeColor;
     }
     if (button == 'graph') {
       graphBtnID.style.color = this.graphColor;